HW7: обрадовался что все написал, оказывается неправильно понял
ну вот на кой там добавление PermissionCheck через Assembly, если все
можно сделать используя уже готовые интерфейсы IAdminRequest и
IClientRequest?

Несколько часов чтобы понять что от меня требуют, наконец-то написал,
идеальное решение, соответствует всем требованиям... увидел Assembly
в PermissionCheck и понял что вам точно нужно получше писать условия.

+++ Dotnet.Homeworks.Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,19 +1,26 @@
-﻿using Dotnet.Homeworks.Infrastructure.Cqrs.Commands;
+﻿using Dotnet.Homeworks.Features.Users.Commands.CreateUser;
+using Dotnet.Homeworks.Infrastructure.Cqrs.Commands;
 using Dotnet.Homeworks.Infrastructure.UnitOfWork;
+using Dotnet.Homeworks.Infrastructure.Validation.Decorators;
+using Dotnet.Homeworks.Infrastructure.Validation.PermissionChecker;
 using Dotnet.Homeworks.Shared.Dto;
+using FluentValidation;
 
 namespace Dotnet.Homeworks.Features.Users.Commands.DeleteUser;
 
-public class DeleteUserCommandHandler: ICommandHandler<DeleteUserCommand>
+public class DeleteUserCommandHandler: CqrsDecorator<DeleteUserCommand, object>, ICommandHandler<DeleteUserCommand>
 {
-    public DeleteUserCommandHandler(UnitOfWork unitOfWork)
+    public DeleteUserCommandHandler(UnitOfWork unitOfWork, IPermissionCheck permissionCheck, IValidator<DeleteUserCommand>? validator): base(permissionCheck, validator)
     {
         UnitOfWork = unitOfWork;
     }
 
     private UnitOfWork UnitOfWork { get; }
-    public async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+    public new async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        var decResult = await base.Handle(request, cancellationToken);
+        if (decResult.IsFailure)
+            return decResult;
         await UnitOfWork.UserRepository.DeleteUserByGuidAsync(request.Guid, cancellationToken);
         await UnitOfWork.SaveChangesAsync(cancellationToken);
         return new Result(true);
